The code has been divided into components for better understandablity
